Move check_auth_blocking to new class AuthBlocking
